Create StaticCardAccountRanges
Summary
The implementation maintains the list of static `AccountRange`s.

Motivation
Allow for changing the static `AccountRange` list in tests.
