+++ stripe/src/main/java/com/stripe/android/cards/StaticCardAccountRanges.kt
@@ -1,0 +1,9 @@
+package com.stripe.android.cards
+
+import com.stripe.android.model.CardMetadata
+
+internal interface StaticCardAccountRanges {
+    fun match(
+        cardNumber: CardNumber.Unvalidated
+    ): CardMetadata.AccountRange?
+}
